refactor: ViewController 간소화를 위한 view 분리
